refactor: streamline workflow loading and validation process
